Cleanups, enable compiling everything as if it were C++.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@28022 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e16/e/src/misc.c
@@ -179,7 +179,7 @@ ETimedLoopNext(void)
 
    EobjsRepaint();
 
-   return y;
+   return (int)y;
 }
 
 /*
